Squashed commit of the following:
commit 8aa2b259ca6656a6b3c98a0c18dc4e686a1803b2
Author: James Popoli <jrpopoli@g.ucla.edu>
Date:   Fri Mar 21 13:23:08 2025 -0700

    Fms implementation (#44)

    Factor match scoring of pf2 components with Andrew's pf2 function
    against resampled disease data.
      - experiments folder for storing wandb experiments and functional code
    - `wandb_pf2_l1_rank.py` file to explore l1 regularization and rank
    effects on fms
    - `figure_pf2_fms.py` to visualize bootstrapping effects on fms and r2x
      - GPU targeting
    - Using a modified parafac2 library to re-enable GPU acceleration.
    `perform_parafac2()` wrapper now targets a specific GPU to avoid job
    overloading
