+++ mrsa_ca_rna/utils.py
@@ -242,3 +242,37 @@ def check_sparsity(array: np.ndarray, threshold: float = 1e-4) -> float:
     A[np.abs(A) < threshold] = 0
     sparsity = 1.0 - (np.count_nonzero(A) / A.size)
     return sparsity
+
+
+def resample_adata(X_in: ad.AnnData) -> ad.AnnData:
+    """Resamples AnnData with unique observation indices, with replacement.
+
+    Parameters
+    ----------
+    X_in : ad.AnnData
+        AnnData object to be resampled
+
+    Returns
+    -------
+    ad.AnnData
+        Resampled AnnData object with unique observation indices
+    """
+
+    # make a random index with replacement for resampling
+    random_index = np.random.randint(0, X_in.shape[0], size=(X_in.shape[0],))
+
+    # independently subset the data and obs with the random indices
+    assert isinstance(X_in.X, np.ndarray)
+    X_resampled = X_in.X[random_index]
+    obs_resampled = X_in.obs.iloc[random_index].copy()
+
+    # Create unique indices for the resampled observations
+    obs_resampled.index = [f"bootstrap_{i}" for i in range(len(obs_resampled))]
+
+    # Create a new AnnData object with the resampled data
+    uns_dict = dict(X_in.uns)
+    X_in_resampled = ad.AnnData(
+        X=X_resampled, obs=obs_resampled, var=X_in.var.copy(), uns=uns_dict
+    )
+
+    return X_in_resampled
